requirements.txt
numpy==1.25.2
opencv_python==4.8.0.76
Pillow==10.0.1
ultralytics==8.0.180
